Reduce padding-bottom of last ul in #tasks
The 40px of padding below each ul in the task list could cause unneeded scrollbars when the task list was within 40px of the bottom of the page. This was especially noticeable with the bulky scrollbars in Windows.  This commit reduces that padding of the last ul to 20px, which matches the padding along the right edge.

As a side effect, style.scss was recompiled without the never-used-in-production `-ms-` prefix for CSS transitions.